Updated client.lua
Added check for engine health, if engine health is lower than 500, the RepairVehicle trigger goes through. This is to prevent people from "repairing" a vehicle and setting engine health lower than what the vehicles engine health was before repairing. Credits to Qinak#4037 for the edit, I just did the PR.